feat: added list, list reverser, sorter for list
[List]
The list will fill the function as all my cariere
options from least recent to most recent which i
can just fill in in a div or iframe.

[List reverser]
I wrote a script in javascript, completely
homebrew, to reverse the list if people want to
see it in a different way.

[sorter for lists]
Added a button at the top of the page which says
if you sorted by MOST recent or LEAST recent